refactor(agent): extract review and handoff gate from execution supervisor
