+++ packages/agent/src/loop/execution-supervisor.ts
@@ -1,6 +1,4 @@
-import type { Tool } from "@openomni/protocol";
 import { Session } from "@openomni/session";
-import { AgentMessenger } from "../agent/communication";
 import { BuiltinAgentRegistry } from "../agent/registry";
 import { TaskManager } from "../task/manager";
 import {
@@ -12,6 +10,13 @@ import {
   buildDependencyGraph,
   completeTaskAndUnblockDependents,
 } from "./execution-graph";
+import {
+  decideFailedStepAction,
+  requestHandoffDocument,
+  reviewTaskResult,
+  rotateAgent,
+  sendReviewFeedback,
+} from "./execution-review";
 import type {
   ChildRunResult,
   DependencyGraph,
@@ -28,7 +33,6 @@ import type {
   ExecutionSupervisorConfigInternal,
   ExecutionSupervisorResult,
   ExecutionSupervisorRuntime,
-  FailureAction,
   FailureDecision,
   RunningTask,
   StepOutcome,
@@ -46,8 +50,6 @@ import { RunWorker } from "./run-worker";
 const DEFAULT_DISPATCH_TIMEOUT_MS = 5 * 60 * 1000;
 const DEFAULT_MAX_SUBAGENT_DEPTH = 3;
 const MAX_REJECTIONS_BEFORE_HANDOFF = 3;
-const DISPATCH_AGENT_ID = "dispatch-supervisor";
-const HANDLE_FAILURE_TOOL = "handle_failure";
 const SUPERVISOR_DECISION_TRIGGER_ID = "dispatch-supervisor-trigger";
 
 export type {
@@ -404,6 +406,7 @@ async function executeDispatchGraph(
           hybridRuntime,
           context,
           dispatchAbortController.signal,
+          runSupervisorToolDecision,
         );
 
         if (failureDecision?.action === "skip") {
@@ -480,6 +483,7 @@ async function executeDispatchGraph(
           state,
           context,
           dispatchAbortController.signal,
+          executeChildRunWithAbort,
         );
         rotateAgent(state, input.objective);
       }
@@ -495,119 +499,6 @@ async function executeDispatchGraph(
     }
     cancelRunningChildren(running, "dispatch_cleanup");
   }
-}
-
-async function decideFailedStepAction(
-  objective: string,
-  state: DispatchTaskState,
-  result: ChildRunResult,
-  hybridRuntime: DispatchHybridRuntime | undefined,
-  context: DispatchContext,
-  abortSignal: AbortSignal,
-): Promise<FailureDecision | undefined> {
-  if (!hybridRuntime || abortSignal.aborted) {
-    return undefined;
-  }
-
-  return requestFailureDecision(
-    objective,
-    {
-      stepId: state.task.id,
-      error: result.error,
-      attempts: state.attempts,
-    },
-    hybridRuntime,
-    context,
-    abortSignal,
-  );
-}
-
-async function requestFailureDecision(
-  objective: string,
-  failedStep: {
-    stepId: string;
-    error: string;
-    attempts: number;
-  },
-  hybridRuntime: DispatchHybridRuntime,
-  context: DispatchContext,
-  abortSignal: AbortSignal,
-): Promise<FailureDecision> {
-  const fallbackDecision: FailureDecision = {
-    action: failedStep.attempts <= 1 ? "retry" : "skip",
-    reasoning:
-      failedStep.attempts <= 1
-        ? "Retry once before skipping"
-        : "Skip after repeated failure",
-  };
-
-  let selectedDecision = fallbackDecision;
-
-  const decisionToolExecutor: ToolExecutor = {
-    async execute(calls: Tool.Call[]): Promise<Tool.Result[]> {
-      return calls.map((call) => {
-        if (call.tool !== HANDLE_FAILURE_TOOL) {
-          return {
-            id: crypto.randomUUID(),
-            toolCallId: call.id,
-            output: `Unsupported supervisor tool: ${call.tool}`,
-            isError: true,
-          };
-        }
-
-        selectedDecision = parseFailureDecision(call.input, fallbackDecision);
-
-        return {
-          id: crypto.randomUUID(),
-          toolCallId: call.id,
-          output: JSON.stringify({ decision: selectedDecision }),
-          isError: false,
-        };
-      });
-    },
-  };
-
-  const prompt = [
-    "Handle failed dispatch step execution.",
-    "Call handle_failure with { action, reasoning }.",
-    `Objective: ${objective}`,
-    `Failed Step: ${JSON.stringify(failedStep)}`,
-    `Options: ${JSON.stringify(["retry", "skip", "replan"])}`,
-  ].join("\n\n");
-
-  await runSupervisorToolDecision(
-    prompt,
-    [HANDLE_FAILURE_TOOL],
-    decisionToolExecutor,
-    hybridRuntime,
-    context,
-    abortSignal,
-  );
-
-  return selectedDecision;
-}
-
-function parseFailureDecision(
-  input: Record<string, unknown>,
-  fallbackDecision: FailureDecision,
-): FailureDecision {
-  const actionRaw = input.action;
-  const reasoningRaw = input.reasoning;
-
-  const action: FailureAction =
-    actionRaw === "retry" || actionRaw === "skip" || actionRaw === "replan"
-      ? actionRaw
-      : fallbackDecision.action;
-
-  const reasoning =
-    typeof reasoningRaw === "string" && reasoningRaw.trim().length > 0
-      ? reasoningRaw.trim()
-      : fallbackDecision.reasoning;
-
-  return {
-    action,
-    reasoning,
-  };
 }
 
 async function runSupervisorToolDecision(
@@ -1090,198 +981,6 @@ async function waitForNextResult(
   } finally {
     cleanup();
   }
-}
-
-async function reviewTaskResult(
-  objective: string,
-  state: DispatchTaskState,
-  result: ChildRunResult,
-  context: DispatchContext,
-): Promise<DispatchReviewDecision> {
-  if (!result.success) {
-    return {
-      decision: "reject",
-      feedback: result.error || "Task execution failed",
-    };
-  }
-
-  if (!context.review) {
-    return defaultReviewDecision(result.summary);
-  }
-
-  try {
-    return await context.review({
-      objective,
-      taskId: state.task.id,
-      agentId: state.agentInstanceId,
-      sessionId: state.sessionId,
-      summary: result.summary,
-      attempt: state.attempts,
-      rejectionStreak: state.rejectionStreak,
-      feedbackHistory: [...state.feedbackHistory],
-    });
-  } catch (error) {
-    return {
-      decision: "reject",
-      feedback: `Review function failed: ${toErrorMessage(error)}`,
-    };
-  }
-}
-
-function defaultReviewDecision(summary: string): DispatchReviewDecision {
-  const normalized = summary.trim().toLowerCase();
-  if (normalized.startsWith("reject:")) {
-    return {
-      decision: "reject",
-      feedback: summary.slice("reject:".length).trim() || "Result rejected",
-    };
-  }
-
-  if (normalized === "reject") {
-    return {
-      decision: "reject",
-      feedback: "Result rejected",
-    };
-  }
-
-  return {
-    decision: "accept",
-  };
-}
-
-async function sendReviewFeedback(
-  state: DispatchTaskState,
-  runId: string,
-  feedback: string,
-): Promise<void> {
-  await AgentMessenger.send({
-    traceId: crypto.randomUUID(),
-    sessionId: state.sessionId,
-    runId: runId || crypto.randomUUID(),
-    fromAgentId: DISPATCH_AGENT_ID,
-    toAgentId: state.agentInstanceId,
-    sentAt: new Date().toISOString(),
-    schemaRef: "dispatch.review.feedback.v1",
-    payload: {
-      type: "review_feedback",
-      taskId: state.task.id,
-      feedback,
-      rejectionCount: state.totalRejections,
-    },
-  }).catch(() => {
-    return undefined;
-  });
-}
-
-async function requestHandoffDocument(
-  objective: string,
-  state: DispatchTaskState,
-  context: DispatchContext,
-  abortSignal: AbortSignal,
-): Promise<string> {
-  if (abortSignal.aborted) {
-    return "Handoff unavailable: dispatch aborted";
-  }
-
-  const agent = BuiltinAgentRegistry.get(state.task.agentType);
-  if (!agent) {
-    return `Handoff unavailable: unknown agent type ${state.task.agentType}`;
-  }
-
-  const handoffSpawnedBy =
-    context.parentTaskId && context.parentRunId && context.parentSessionId
-      ? {
-          taskId: context.parentTaskId,
-          runId: context.parentRunId,
-          sessionId: context.parentSessionId,
-        }
-      : undefined;
-
-  const triggerResult = await TaskManager.trigger(state.childTaskId, {
-    triggerId: "dispatch-trigger",
-    type: "manual",
-    occurredAt: Date.now(),
-    spawnedBy: handoffSpawnedBy,
-  });
-
-  if ("error" in triggerResult) {
-    return `Handoff unavailable: failed to create handoff run (${triggerResult.error})`;
-  }
-
-  const prompt = [
-    "Create a handoff document for replacement agent.",
-    `Objective: ${objective}`,
-    `Task ID: ${state.task.id}`,
-    `Task Description: ${state.task.description}`,
-    `Attempt Count: ${state.attempts}`,
-    `Feedback History:\n- ${state.feedbackHistory.join("\n- ") || "none"}`,
-    "Include attempted approach, blockers, and recommended next steps.",
-  ].join("\n\n");
-
-  const config: OrchestratorConfig = {
-    taskId: state.childTaskId,
-    runId: triggerResult.runId,
-    maxRetries: 0,
-    sessionMode: "reuse",
-    sessionId: state.sessionId,
-    maxSubagentDepth: context.maxDepth ?? DEFAULT_MAX_SUBAGENT_DEPTH,
-    currentDepth: (context.parentDepth ?? 0) + 1,
-    insideDelegation: true,
-  };
-
-  const orchestratorInput: OrchestratorRunInput = {
-    llm: context.llm,
-    input: {
-      systemPrompt: agent.systemPrompt,
-      prompt,
-      agentType: agent.name,
-      tools: agent.tools,
-      permissions: agent.permissions,
-      maxTurns: agent.maxTurns,
-    },
-    toolExecutor: context.toolExecutor,
-  };
-
-  const result = await executeChildRunWithAbort(
-    config,
-    orchestratorInput,
-    abortSignal,
-  );
-
-  if (!result.success || !result.summary.trim()) {
-    return [
-      "Handoff summary unavailable from agent.",
-      `Latest error: ${result.error || "none"}`,
-      `Last summary: ${state.summaries[state.summaries.length - 1] || "none"}`,
-      `Feedback:\n- ${state.feedbackHistory.join("\n- ") || "none"}`,
-    ].join("\n\n");
-  }
-
-  return result.summary;
-}
-
-function rotateAgent(state: DispatchTaskState, objective: string): void {
-  const agent = BuiltinAgentRegistry.get(state.task.agentType);
-  if (!agent) {
-    state.errors.push(
-      `Unable to rotate agent: ${state.task.agentType} not found`,
-    );
-    state.status = "failed";
-    return;
-  }
-
-  state.handoffs += 1;
-  state.rejectionStreak = 0;
-  state.agentInstanceId = createAgentInstanceId(state.task.id);
-  state.agentHistory.push(state.agentInstanceId);
-
-  state.sessionId = createSessionId(state.task.id);
-  ensurePersistentSession(
-    state.sessionId,
-    `${objective}: ${state.task.id} (handoff)`,
-    agent.name,
-  );
-  state.childTaskId = createChildTask(state.task, state.agentInstanceId);
 }
 
 function normalizeFileScope(fileScope: string[]): string[] {
